+++ core/src/com/angian/libgdx/pipes/LevelLayout.java
@@ -1,5 +1,6 @@
 package com.angian.libgdx.pipes;
 
+import com.badlogic.gdx.graphics.Color;
 import com.badlogic.gdx.math.Rectangle;
 import com.badlogic.gdx.utils.Align;
 
@@ -7,10 +8,11 @@ public class LevelLayout {
     public static final int SCREEN_WIDTH = 1200;
     public static final int SCREEN_HEIGHT = 900;
 
+    public static final int TILE_SIZE = 100;
+
     public static final Rectangle BOARD = new Rectangle(180, 180, 1000, 700);
 
     public static final int TEXT_FONT_SIZE = 32;
-    //public static final int TEXT_FONT_SIZE = 72;
     public static final int TEXT_Y = 100;
     public static final int TEXT_HEIGHT = 70;
 
@@ -20,10 +22,14 @@ public class LevelLayout {
     public static final Rectangle TEXT_SCORE = new Rectangle(180, TEXT_Y, 1000, TEXT_HEIGHT);
     public static final int TEXT_SCORE_ALIGN = Align.center;
 
-    public static final Rectangle TEXT_DISTANCE = new Rectangle(940, TEXT_Y, 200, TEXT_HEIGHT);
+    public static final Rectangle TEXT_DISTANCE = new Rectangle(960, TEXT_Y, 200, TEXT_HEIGHT);
     public static final int TEXT_DISTANCE_ALIGN = Align.right;
 
-    //TODO: tile preview, timer
+    public static final Rectangle TILE_PREVIEW = new Rectangle(10, 180, 110, 600);
+    public static final int TILE_PREVIEW_PADDING = 20;
+
+    public static final Rectangle TIMER = new Rectangle(135, 380, 20, 500);
+    public static final Color TIMER_COLOR = Color.ORANGE;
 
     public static Rectangle standard2gdxCoords(Rectangle layoutRect) {
         float flippedY = SCREEN_HEIGHT - layoutRect.y -layoutRect.height;