+++ core/src/com/angian/libgdx/pipes/PipesGame.java
@@ -10,6 +10,6 @@ public class PipesGame extends BaseGame {
 	public void create () {
 		super.create();
 
-		setActiveScreen(new LevelScreen());
+		setActiveScreen(new MenuScreen());
 	}
 }