+++ CHANGELOG.md
@@ -6,7 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ## [Unreleased] - yyyy-mm-dd
 ...
 
-## [0.2] - 2022-01-14
+## [0.2] - 2022-01-17
+### Added
+- preview tiles
+- full basic artwork
+- water timer
+- start screen
+
+
+## [0.2] - 2022-01-15
 
 ### Added
 - screen background